% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_utils.R
\name{ad_3d_array_to_snp_index_mat}
\alias{ad_3d_array_to_snp_index_mat}
\title{Get SNP-Index matrix from AD 3D array.}
\usage{
ad_3d_array_to_snp_index_mat(ad_3d_array)
}
\description{
Get SNP-Index matrix from AD 3D array.
}
