% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtf_utils.R
\name{add_biotype_by_cds}
\alias{add_biotype_by_cds}
\title{Add biotype for VEP annotation.}
\usage{
add_biotype_by_cds(
  gtf,
  cds_type = "CDS",
  coding_biotype = "protein_coding",
  noncoding_biotype = "ncRNA"
)
}
\arguments{
\item{gtf}{a GTF GRanges.}

\item{cds_type}{CDS type string. Default is \code{"CDS"}.}

\item{coding_biotype}{biotype string for coding genes and transcripts. Default is \code{"protein_coding"}}

\item{noncoding_biotype}{biotype string for non-coding genes and transcripts. Default is \code{"ncRNA"}}
}
\description{
Add biotype for VEP annotation.
}
