% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimap_homolog.R
\name{aln_to_dna}
\alias{aln_to_dna}
\title{Fetch target sequence from target genome by alignment.}
\usage{
aln_to_dna(aln, target_genome)
}
\arguments{
\item{aln}{alignment object generated by \code{\link[=minimap_homolog]{minimap_homolog()}}.}

\item{target_genome}{target genome FASTA path.}
}
\description{
Fetch target sequence from target genome by alignment.
}
