% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_utils.R
\name{bcftools_variant_stats}
\alias{bcftools_variant_stats}
\title{Calculate MAF and Missing rate of variants in vcf df
using bcftools +fill-tags}
\usage{
bcftools_variant_stats(vcf)
}
\arguments{
\item{vcf}{a vcf tibble or vcf file path.}
}
\description{
Calculate MAF and Missing rate of variants in vcf df
using bcftools +fill-tags
}
