% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype.R
\name{beagle_wrapper}
\alias{beagle_wrapper}
\title{Wrapper for running beagle to impute vcf}
\usage{
beagle_wrapper(gt, out, nthreads = parallel::detectCores(), seed = 12345)
}
\arguments{
\item{gt}{path to .vcf.gz file.}

\item{out}{output prefix, output file will be \code{prefix.vcf.gz}.}

\item{nthreads}{number of threads beagle can use.}
}
\description{
Wrapper for running beagle to impute vcf
}
