% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cigar_liftover.R
\name{cigar_liftover}
\alias{cigar_liftover}
\title{Use CIGAR string to shift position}
\usage{
cigar_liftover(pos_vect, query_start, aln)
}
\arguments{
\item{pos_vect}{Positions that need to be shifted.}

\item{query_start}{Start position of query sequence in query genome (1-based index).}

\item{aln}{A GAlignments object (from GenomicAlignments) contains CIGAR field.}
}
\value{
A integer vector of new position match \code{pos_vect}.
}
\description{
Use CIGAR string to shift position
}
\details{
First, find matched windows between query sequence and target sequence.
Then, check each window pair, shift pos from query to target.
}
