% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{geom_anova_lsd_groups}
\alias{geom_anova_lsd_groups}
\title{Add signif stars (ANOVA LSD)}
\usage{
geom_anova_lsd_groups(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  label_y_fun = label_y_fun_max_0.1,
  ...
)
}
\arguments{
\item{label_y_fun}{A function for signif star y pos.}
}
\description{
Perform ANOVA and LSD tests, then plot LSD letters.
}
