% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{geom_test_with_ref}
\alias{geom_test_with_ref}
\title{Add signif stars (test with ref)}
\usage{
geom_test_with_ref(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  label_y_fun = label_y_fun_max_0.1,
  ref_group = 1,
  test = "t.test",
  ...
)
}
\arguments{
\item{label_y_fun}{A function for signif star y pos.}

\item{ref_group}{Index of reference group level.}

\item{test}{Name of function for run test.}
}
\description{
Test all groups with ref group and plot signif stars.
}
