% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{geom_triad}
\alias{geom_triad}
\title{Plot bar, errorbar and jitter}
\usage{
geom_triad(bar_args = NULL, errorbar_args = NULL, jitter_args = NULL)
}
\arguments{
\item{bar_args}{Arguments for barplot.}

\item{errorbar_args}{Arguments for errorbar.}

\item{jitter_args}{Arguments for jitter.}
}
\description{
Plot bar, errorbar and jitter dots in one function call with proper default
settings.
}
