% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_track.R
\name{get_gene_range}
\alias{get_gene_range}
\title{Get Gene Nearby Range}
\usage{
get_gene_range(gtf, gene, upstream = 2000, downstream = 2000)
}
\arguments{
\item{gtf}{a GRanges imported by rtracklayer from GTF file.}

\item{gene}{name of a gene.}

\item{upstream}{upstream bp.}

\item{downstream}{downstream bp.}
}
\description{
Given a gene name, return its nearby region GRanges.
}
