% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtf_utils.R
\name{gff_to_gtf}
\alias{gff_to_gtf}
\title{Convert GFF GRanges to GTF GRanges.}
\usage{
gff_to_gtf(gff, gene_type = "gene", transcript_type = "mRNA")
}
\arguments{
\item{gff}{a GRanges object read by \code{\link[rtracklayer:import]{rtracklayer::import()}}}

\item{gene_type}{string for gene feature type. Default is \code{"gene"}.}

\item{transcript_type}{string for transcript feature type. Default is \code{"mRNA"}.}
}
\description{
Convert GFF GRanges to GTF GRanges.
}
