% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{group_info_k9}
\alias{group_info_k9}
\title{K9 group of 3K-RG samples}
\format{
An object of class \code{factor} of length 3023.
}
\source{
\url{https://v1.rmbreeding.cn/tool/3kgroup}
}
\usage{
group_info_k9
}
\description{
A named factor, names are sample names, values are K9 group info.
}
\keyword{datasets}
