% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype.R
\name{hap_long_to_hap_wide}
\alias{hap_long_to_hap_wide}
\title{Convert hap_long to hap_wide}
\usage{
hap_long_to_hap_wide(hap_long)
}
\arguments{
\item{hap_long}{generated by hap_mat_to_hap_long.}
}
\description{
Convert hap_long to hap_wide
}
