% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype.R
\name{hap_mat_to_hap_long}
\alias{hap_mat_to_hap_long}
\title{Count haplotypes per group}
\usage{
hap_mat_to_hap_long(hap_mat, group_info)
}
\arguments{
\item{hap_mat}{haplotype matrix produced by \code{vcf_to_hap_mat}.}

\item{group_info}{a named vector or factor that indicates each haplotype's group.}
}
\description{
Count haplotypes per group
}
