% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype.R
\name{hap_wide_remove_consensus}
\alias{hap_wide_remove_consensus}
\title{After filtering hap_wide, remove consensus bases in Seq.}
\usage{
hap_wide_remove_consensus(hap_wide_subset)
}
\arguments{
\item{hap_wide_subset}{a subset of hap_wide (require "pos" attribute retained).}
}
\description{
After filtering hap_wide, remove consensus bases in Seq.
}
