% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_wise_intersect.R
\name{k_wise_intersect}
\alias{k_wise_intersect}
\title{K-wise intersection of GRanges}
\usage{
k_wise_intersect(gr_list, k)
}
\arguments{
\item{gr_list}{a list of GRanges}

\item{k}{e.g. a 2-wise intersection of 3 GRanges means 3 choose 2.}
}
\description{
K-wise intersection of GRanges
}
