% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cigar_liftover.R
\name{parse_cigar}
\alias{parse_cigar}
\title{Convert CIGAR string to named character vector}
\usage{
parse_cigar(cigar)
}
\arguments{
\item{cigar}{A CIGAR string from SAM file. e.g. \code{c("2M1I6M")}}
}
\value{
a named character vector. e.g. \code{c(M=2, I=1, M=6)}
}
\description{
Convert CIGAR string to named character vector
}
