% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_track.R
\name{plot_gene_track}
\alias{plot_gene_track}
\title{Plot Gene Track with Primary Transcripts}
\usage{
plot_gene_track(
  gtf,
  gene_range,
  padding = 0,
  show_gene_id = TRUE,
  show_arrow = TRUE,
  remove = FALSE,
  keep_only = FALSE
)
}
\arguments{
\item{gtf}{A GRanges, generated by rtracklayer::import from a GTF file in ensembl format.}

\item{gene_range}{A GRanges, specify a region of gtf to plot gene track.}

\item{padding}{A integer, it is the padding of genes' upstream and downstream so that its name not overlapping.}

\item{show_gene_id}{Boolean, Whether to show gene_id of transcripts.}

\item{show_arrow}{Boolean, Whether to show arrow indicating orientation of transcripts.}

\item{remove}{A vector of gene_id, remove transcripts indicated.}

\item{keep_only}{A vector of gene_id, keep transcripts indicated.}
}
\description{
Plot Gene Track with Primary Transcripts
}
