% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_utils.R
\name{vcf_to_gt_mat}
\alias{vcf_to_gt_mat}
\title{Get genotype character matrix from vcf tibble.}
\usage{
vcf_to_gt_mat(vcf_df, simplify = FALSE)
}
\arguments{
\item{vcf_df}{vcf tibble or data.frame.}

\item{simplify}{e.g. "T/T" > "T", "AG/AG" > "AG", won't simplify heterozygote.}
}
\description{
Get genotype character matrix from vcf tibble.
}
