% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype.R
\name{vcf_to_hap_mat}
\alias{vcf_to_hap_mat}
\title{Convert vcf df to haplotype matrix}
\usage{
vcf_to_hap_mat(vcf)
}
\arguments{
\item{vcf}{vcf data.frame or tibble.}
}
\description{
Row is haplotype, column is variant.
}
