% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_utils.R
\name{vcf_variant_stats_R}
\alias{vcf_variant_stats_R}
\title{Calculate MAF and Missing rate of variants in vcf df
inside R}
\usage{
vcf_variant_stats_R(vcf)
}
\arguments{
\item{vcf}{a vcf tibble.}
}
\description{
Calculate MAF and Missing rate of variants in vcf df
inside R
}
