% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bed.R
\name{write_broadPeak}
\alias{write_broadPeak}
\title{Export Granges as narrowPeak file}
\usage{
write_broadPeak(gr, file)
}
\arguments{
\item{gr}{GRanges imported from a narrowPeak file using \code{\link[rtracklayer:import]{rtracklayer::import()}}}

\item{file}{path to output file}
}
\description{
Export Granges as narrowPeak file
}
