% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_narrowPeak.R
\name{write_narrowPeak}
\alias{write_narrowPeak}
\title{Export Granges as narrowPeak file}
\usage{
write_narrowPeak(gr, file)
}
\arguments{
\item{gr}{GRanges imported from a narrowPeak file}

\item{file}{path to output file}
}
\description{
Export Granges as narrowPeak file
}
