% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype.R
\name{write_nexus}
\alias{write_nexus}
\title{Write hap_wide df to nexus file.}
\usage{
write_nexus(hap_wide, out)
}
\arguments{
\item{hap_wide}{a hap_wide object.}

\item{out}{Path to output file.}
}
\description{
Write hap_wide df to nexus file.
}
