% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_utils.R
\name{write_vcf_bgzip}
\alias{write_vcf_bgzip}
\title{Write VCF tibble as bgzipped file.}
\usage{
write_vcf_bgzip(vcf, file)
}
\arguments{
\item{vcf}{A VCF tibble.}

\item{file}{Path to output bgzipped file.}
}
\description{
Write VCF tibble as bgzipped file.
}
