% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_utils.R
\name{ad_string_mat_to_ad_3d_array}
\alias{ad_string_mat_to_ad_3d_array}
\title{Parse AD string matrix to a 3D Array.}
\usage{
ad_string_mat_to_ad_3d_array(ad_string_mat)
}
\arguments{
\item{ad_string_mat}{a matrix generated by \code{vcf_to_field_matrix(vcf, "AD")}.}
}
\description{
Parse AD string matrix to a 3D Array.
}
