% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimap_homolog.R
\name{aln_to_genbank}
\alias{aln_to_genbank}
\title{Generate genbank string from target genome by alignment.}
\usage{
aln_to_genbank(aln, target_genome, target_gtf)
}
\arguments{
\item{aln}{alignment object generated by \code{\link[=minimap_homolog]{minimap_homolog()}}.}

\item{target_genome}{target genome FASTA path.}

\item{target_gtf}{target genome annotation GTF path.}

\item{preset}{a string, which is the preset of minimap2.}
}
\description{
Generate genbank string from target genome by alignment.
}
