% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimap_homolog.R
\name{minimap_homolog}
\alias{minimap_homolog}
\title{Use external minimap2 to find homolog in another genome.}
\usage{
minimap_homolog(target, query, chromosome, start, end, preset = NULL)
}
\arguments{
\item{target}{target genome FASTA file path.}

\item{query}{query sequence. Either a path to a FASTA file or a \code{\link[Biostrings:DNAStringSet]{Biostrings::DNAStringSet()}} object..}

\item{chromosome, start, end}{position of segment in query sequence. Optional, leave blank to search the entire query sequence.}
}
\description{
Use external minimap2 to find homolog in another genome.
}
