% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterpie_map.R
\name{plot_scatter_pie_map}
\alias{plot_scatter_pie_map}
\title{Plot scatter pie map.}
\usage{
plot_scatter_pie_map(plot_data, sf_obj)
}
\arguments{
\item{plot_data}{a long data.frame, contains these columns: SampleID, HapGroup, lon, lat.}

\item{sf_obj}{a sf object read from a shapefile by \code{\link[sf:st_read]{sf::read_sf()}}.}
}
\value{
a ggplot plot object.
}
\description{
Plot scatter pie map.
}
