% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_utils.R
\name{vcf_to_field_matrix}
\alias{vcf_to_field_matrix}
\title{Get field matrix from vcf tibble.}
\usage{
vcf_to_field_matrix(vcf, field, mc.cores = NULL)
}
\arguments{
\item{vcf}{a vcf tibble.}

\item{field}{field name, e.g. "AD", "DP"}

\item{mc.cores}{number of cores used in parallel::mclapply}
}
\description{
Get genotype field sample columns of vcf tibble, such as AD, DP...
}
